use client for aas search with injected repoClient
