+++ src/app/[locale]/viewer/[base64AasId]/page.tsx
@@ -13,12 +13,12 @@ import { showError } from 'lib/util/ErrorHandlerUtil';
 import { AssetAdministrationShell, LangStringNameType, Reference } from '@aas-core-works/aas-core3.0-typescript/types';
 import { useIsMobile } from 'lib/hooks/UseBreakpoints';
 import { getTranslationText } from 'lib/util/SubmodelResolverUtil';
-import { useRouter, useParams } from 'next/navigation';
+import { useParams, useRouter } from 'next/navigation';
 import { SubmodelsOverviewCard } from '../_components/SubmodelsOverviewCard';
 import { AASOverviewCard } from 'app/[locale]/viewer/_components/AASOverviewCard';
 import { useApis } from 'components/azureAuthentication/ApiProvider';
 import { useEnv } from 'app/env/provider';
-import { handleAasRegistrySearch } from 'lib/searchUtilActions/search';
+import { handleAasRegistrySearch } from 'lib/searchUtilActions/searchServer';
 
 export default function Page() {
     const navigate = useRouter();
